# Conflicts:
#	.dart_tool/package_config_subset
#	.idea/libraries/Flutter_Plugins.xml
#	.idea/workspace.xml
#	android/.gradle/6.7/executionHistory/executionHistory.bin
#	android/.gradle/6.7/executionHistory/executionHistory.lock
#	android/.gradle/6.7/fileHashes/fileHashes.bin
#	android/.gradle/6.7/fileHashes/fileHashes.lock
#	android/.gradle/6.7/fileHashes/resourceHashesCache.bin
#	android/.gradle/6.7/javaCompile/classAnalysis.bin
#	android/.gradle/6.7/javaCompile/jarAnalysis.bin
#	android/.gradle/6.7/javaCompile/javaCompile.lock
#	android/.gradle/6.7/javaCompile/taskHistory.bin
#	android/.gradle/buildOutputCleanup/buildOutputCleanup.lock
#	android/.gradle/buildOutputCleanup/outputFiles.bin
#	android/.gradle/checksums/checksums.lock
#	android/.gradle/checksums/md5-checksums.bin
#	android/.gradle/checksums/sha1-checksums.bin
#	android/local.properties

+++ lib/Forum/my_forum.dart
@@ -3,6 +3,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 
 import 'package:get/get.dart';
+import 'package:velocity_x/velocity_x.dart';
 
 class MyForums extends StatelessWidget {
   FirebaseFirestore _firestore = FirebaseFirestore.instance;
@@ -52,10 +53,18 @@ class MyForums extends StatelessWidget {
                             color: Colors.white,
                             onPressed: () {
                               print(forumData.id);
-                              _firestore
-                                  .collection('forums')
-                                  .doc(forumData.id)
-                                  .delete();
+                              VxDialog.showConfirmation(context,
+                                  title: 'Are you sure?',
+                                  content:
+                                      'Once deleted you will not able to recover this data',
+                                  confirm: 'Delete It',
+                                  confirmBgColor: Colors.redAccent,
+                                  cancel: 'No', onConfirmPress: () {
+                                _firestore
+                                    .collection('forums')
+                                    .doc(forumData.id)
+                                    .delete();
+                              });
                             },
                           ),
                         ),
